refactor how we enforce write permissions for the recording directory
instead of a start-service time check, this is now enforced by the type
system and will fail when the config is being loaded if it isn't
correct.  if we find that users who don't want to use the file
recordings feature are bothered by the fact that they must provide a
writeable directory for them, we can soften the severity of the failure.