call stopService() if the service fails to load, so we send an idle presence
in the error case, the service was calling stop on itself but
jibrimanager did not call its 'stopService' so we were failing to clean
up the state.  this is a quick fix that solves the problem, but i plan
to come back and clean up this mechanism so it's cleaner.