fix an issue with restarting ffmpeg (without adding another bug this (#102)
time)

was shutting down the executor, which means it will reject all future
tasks we try to submit to it.  since we use the same ffmpegexecutor
instance on restart, this was breaking restarting ffmpeg.  Also make
sure we let Tee run to its natural completion, so we get a proper
closing of all of its downstream branches.