Jibri no longer deletes/moves files after finalize failure/success
Jibri will still execute the finalize script, passing the directory in
which the most recent recording has been written, but it will no longer
be taking responsibility for the recorded files after that point
(specifically, it will not check the return code of finalize--only log
it--and will not delete successfully finalized recordings nor move
unsuccessfully finalized ones).

This is because the new recommendation will be to have a separate
process handle the finished/failed recordings so that we don't block the
jibri while it has to do that work.