better handling of ffmpeg launch error in FfmpegExecutor
the creation of a processwrapper was wrapped in a try/catch, but it's
actually #start() that potentially throws (i think previously
ProcessWrapper launched the command at construction).  if ffmpeg wasn't
installed, for example, start would throw which was fine, as we handled
the exception and started shutting things down.  but in
ProcessWrapper#stop we assume the process was correctly created, so when
we try to stop it we threw again (lateinit var wasn't initialized).
enforce now in FfmpegExecutor that the process variable is only held if
it is in a valid state.