fix issue when we stop due to a call status check
we performed the statemachine transition from within the recurring call
check, which bubbled up a status change and eventually lead to a call
to leaveCallAndQuitBrowswer all within the same context we performed the
call status check in.  since we cancel the recurring call check task
in leaveCallAndQuitBrowser, we tried cancelling the task from within the
context of which the task was running.

another option would've been to cancel the task but not interrupt it,
but it felt wrong to have all of this execute from within the recurring
task's context, so instead we'll post the state machine transition to
another pool from within the recurring task.