fix: webhooks JWT recreation after expiry
Previous fix has slightly refactored the handling of the JWT header,
thereby introducing a regression.
When passing the defaultRequest closure we must ensure
we supply a block that forces re-evaluation of the `jwt` property,
thereby executing the expiry check in RefreshingProperty.
As such, `$jwt` is valid, while `$it` is not, because `it` is resolved
only once, when the Client is setup. Whatever value is computed then,
is seen as a constant when passed to defaultRequest.
`$jwt`, on the other hand, forces re-evaluation every time the default request
block is executed, i.e. for every webhook request.

https://community.jitsi.org/t/jibri-is-not-sending-webhook-authorization-header/117184/